Drop generated identity values for static data
The pay element types and the pay bands are static data which we'll
need to be able to refer from other systems so using generated values
doesn't make sense. For example, the listener service will need to know
the id for the relevant pay element type to create a productive entry
in the pay elements table. I suggest we scope the types into three
bandings based on current requirements:

1xx: Non-productive types such as dayworks and annual leave
2xx: Adjustments such as manual adjustments and balance brought forward
3xx: Productive types - initially reactive repairs but can be expanded

Similarly, the schemes and pay bands once setup should not change so
I suggest the following numbering:

1: SMV scheme
2: Electrician
3: Plumber
4: Carpenter
5: Glazier
6: Gas Engineer
7: Roofer
8: Legacy SMV

The pay band ids will fall naturally then into ranges such as 11-19 for
the SMV scheme, 21-29 for the Electrician scheme and so on. Note that
not all of the schemes may be in use but these bandings were identified
based on all available data including the last 0093.pdf report. In that
report it seems that some operatives were on an SMV scheme that had
slightly lower values for the bands 2-6 (5% lower, e.g. 140% for band 3
vs. 145% for band 3 reported in other documentation).