Switch to a real database for gateway tests
We want to use views for the operative summaries and the in-memory db
doesn't support this. Also switch to using Migrate over EnsureCreated
so that migrations for views are applied in the test database.

Another reason to switch to using a real database is that it's something
of a contentious point within the EF Core community as demonstrated by
the discussion on the following GitHub issue:

https://github.com/dotnet/efcore/issues/18457