Add a selectable flag to pay element types
This is to control what appears in the drop down boxes when pay elements
are being edited in the frontend application.

Additionally marking adjustment pay elements as productive so that they
appear in the summary reports. This allows us to use the 200-299 range
for the 'Productive not in Repairs Hub' feature and keep the 300+ range
for productive time coming from Repairs Hub.