Add a max_value to a scheme
Operatives on band 9 don't carry over SMVs to the next period so we need
to cap their SMVs earned. Adding this value will allow use to use LEAST to
operate that cap without having to calculate it on the fly by doing a
second join to the pay_bands table on band 9 and multiplying by 13 weeks.