Add generated pay band columns to ease calculations
To calculate the carry over into the next period we need to know how much
an operative has exceeded the band lower limit by so use a generated column
to ease that calculation and avoid have to have `value * 13` everywhere.

Similarly, once band changes have been applied we need to recalculate the
non-productive SMVs for the new period so to aid in that calculate the SMVs
per hour on each band so we can directly multiply against the duration on
the pay element record and not have to have `value / 36` everywhere.