Add BandChange entity and associated views
The band_changes table is to be populated from the operative_projections view
when the last week of the period is closed. We need to snapshot things all the
operative data that might change to keep an accurate record of what happened.

Note that the operative_projections table which is the source for the initial
band change records after the 13th week closes uses the projected band for the
band_value but for the carry overs to be correct it needs to use the final_band
value once the band change process has completed.