Push down coalesce for cost code to elements
Some pay elements have the operative's default code code assigned
directly and by using a coalesce at the summary view we can end up
with two entries for the same operative and cost code. This confuses
EF as it's expecting the two columns to be a primary key and so
leads to duplicate rows in the downloaded CSV reports.

Fix by pushing the coalesce into the elements view which will add
the overhead of the additional two table joins to get the section
column from the operative record but it should only ever be a few
hundred records at a time and as a consequence a small portion
of the total request processing time.