Remove ThrowHelper as it confuses SonarCloud
The ThrowHelper class masks from SonarCloud the fact that an exception is
being raised so either gives a false warning about a null object reference
or a warning about a code path not returning a value. It's less effort to
write the exception throwing code out in full than trying to work around
the false positive generated by SonarCloud so remove the helper class.