Add an endpoint for getting a bonus period
We finally have a need for a get period endpoint - Brian needs to see the
band changes after the period has been closed so we need to fetch the
band period in question. The band changes already has an endpoint which
was added for doing the previous payroll report feature.